requirements.txt
flask>=3.0.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
pillow>=10.0.0
scipy>=1.12.0
